\name{miha_run}
\alias{miha_run}
\title{Convert wheel rotations to distance}
\usage{
miha_run(rotations)
}
\arguments{
  \item{rotations}{Number of wheel rotations}
}
\value{
A vector containing distance in miles and kilometers
}
\description{
Converts the number of wheel rotations to distance measurements
in both miles and kilometers.
}
\examples{
miha_run(10)
}